+++ lib/arel_extensions/nodes/aggregate_function.rb
@@ -4,8 +4,8 @@ module ArelExtensions
       attr_accessor :order, :group
 
       def initialize node, **opts
-        @order = Array(opts[:order]).map{|e| convert_to_node(e)}
-        @group = Array(opts[:group]).map{|e| convert_to_node(e)}
+        @order = Array.wrap(opts[:order]).map{|e| convert_to_node(e)}
+        @group = Array.wrap(opts[:group]).map{|e| convert_to_node(e)}
         super [node]
       end
     end