+++ NEWS.md
@@ -3,6 +3,7 @@
 ## [unreleased]
 
 - Completely removed `==` and `!=` on attributes. Use `.eq` and `.not_eq` instead.
+- Fix a bug with group concat that didn't correctly check for the separator value.
 
 ## Release v2.2.2/v1.4.2 (27-11-2024)