+++ NEWS.md
@@ -3,6 +3,7 @@
 ## [unreleased]
 
 - Use Ruby's Warning categories for Ruby > 2 for deprecations warnings.
+- Fix a bug with group concat that didn't correctly check for the separator value.
 
 ## Release v2.2.2/v1.4.2 (27-11-2024)